fixes for Eval Pane5's TestInput
Upstream pull request #334 pointed out that the Test Input pane in
ScreenEvaluation was positioned incorrectly when playing double as
PLAYER_2.

I went to investigate and found that the entire pane was throwing errors
and failing to load in this branch.  I fixed that by removing the
Name="Pane5" attribute, since the InputHandler looks for unnamed
ActorFrames when cycling through panes.

With that in place, the pad graphics were indeed offset too much, so I
brought over Valentina16's changes from #334.