There's a lot of songs on cabby now, Hoping to make navigation a little easier.
This is handy when you're playing packs with series. For example, often players want to play an "Easy as Pie" song rather than specifically an "Easy as Pie 4" song and tend to look through each song in that series. This allows the player to quickly navigate up and down through packs like Easy as Pie, Piece of Cake, Dbk, Valex, Cringle, Anime Extravaganza, etc.

(Based on observations at Cabby)