Don't force "loading" of local high scores
The current implementation displays ??.?? or ---- while it is loading scores from groovestats. The problem however, is that it does this even when there are scores present on the local machine.

For example: let's say in 15guys1pack there's a song called "Encoder - Future Bass Remix" and the following are true:
	- There is a full leaderboard on our local machine
	- The top score on our machine is 100%
	- There is an extensive leaderboard on Groovestats
	- The top score on groovestats is 98.65%
	- Your profile's top score is a 85% and this information is saved both locally on your machine/USB and on Groovestats

When the player hovers over "Encoder - Future Bass Remix" to view the top score, it will show ??.??% --- while it checks Groovestats and after a little bit of time (depending on your internet connection) it will update with the Machine High Score. It will also show ??.??% ---- for your personal high score while it waits on groovestats

This is less of a problem in this instance, in a way it *makes sense* why this happens. It becomes more of an issue when we enter a situation where someone is hovering over a song that is not ranked on Groovestats. Even though this song is not ranked on groovestats we still have to wait for it to check groovestats before showing us the top machine score. This problem makes it impossible to rapidly scroll through songs in search of score (instead we have to hover over a song for 1-8s while it checks groovestats)

In this commit, we return default functionality to score displaying and the value will then be updated if there is a score on Groovestats. That means that it will show whatever score is on the machine outright without having to load first, and if there is a score on groovestats, you can hover over the song and wait for groovestats to load and update the value.

The next step now is to find a way to properly distinguish Machine High Scores from Groovestats high scores.