Enhancements of Select Profile II
The goal is for cabby to use NFC readers and/or card swipes to choose profiles rather than the profile selection screen. It has caused issues in the past few years. There's heavy inspiration in the current design from DDR.

Notable TO-DOs  - Make the demo change based on readspeed and mini