hide stepartist, bouncing cursors in SSM
As my inline comment states, the new UI for autosetstyle uses tabs along the top
of each player's PaneDisplay in ScreenSelectMusic, and with this change, there's
no longer space (or a need) for the grid of difficulty blocks.  And with the
grid of difficulty blocks gone, there's no longer a need for bouncing cursors or
an obvious place to stick stepartist text.

For now, don't load either of those if AutoSetStyle is set to true under
[Common] in Metrics.ini