prevent crash when backing out of Gameplay via ESC
I'd previously added this code to handle a UX quirk with AutoSetStyle (now
described via inline comment) and playing more than 1 song in a set.

When this code was previously in ./Scripts/SL-Branches.lua, it was being called
in that murky engine area between one screen being done and ScreenSelectMusic
not being ready yet.

I found that holding Escape to back out of gameplay with this code in place was
causing StepMania to crash.  Moving it ScreenSelectMusic's lua files prevented
the crash and still fixed the UX quirk, so we'll go with that.

I've added inline comments walking through the issue and my reasoning.