Add Pad codes for favoriting songs, Favorite songs on Evaluation Screen
Right now, it's kind of tedious to favorite a lot of songs. Especially on a public cab where you can't necessarily just edit the txt file.
You have to open the menu hit favorite and do this for every song. Additionally if you're on a public cab you have to insert more money
before you can favorite a song because you can only favorite them in the MusicWheel (So if I just played a song and I'm like WOW that was GREAT, I wanna add that to my favorites so I don't forget!
I have to pay again and get to the music wheel before I can favorite it)

This adds a pad code to favorite songs. You can use it in the MusicWheel or on the Evaluation Screen. We've grown quite used to this at cabby over the years
and it's made favoriting a lot more convenient. The pad code is a "spin" technically so it's not something that would really trigger accidentally and it can be done quickly.
Now you can favorite songs and scroll through the musicwheel to favorite a bunch of things and you can favorite a song after you've just played it while lingering on evaluation.

In pump the favorite code is the same as unlocking Full Mode on a normal pump machine. Audio uses the unlock sounds that were previously in SL.
(cherry picked from commit eb29744826e1f872f09d83fd5ebd37704ae1d417)