Add a new Theme Preference to toggle Density Graph and Tech display in ScreenSelectMusic
This commit adds a new preference called "SelectMusicDisplayStyle" which can toggle the "Classic" UI for the selectmusic screen.

The Chart Parser allows SL to show the Density Graph and any tech detected in a given chart. This is pretty useful, especially for advanced players. However, I found that this was a bit intimidating to new players. It also takes a way a large part of the sight reading experience. This commit allows a user to toggle whether they want to show this additional information or not.