ensure Casual Mode respects #SELECTABLE:NO;
This fixes CasualMode to not show songs that are considered "locked" by
StepMania.  This includes one or more of:

  #SELECTABLE:NO is set in the simfile and StepMania's HiddenSongs preference is
    enabled.
  the song is locked using UNLOCKMAN
  the song is disabled via ScreenOptionsToggleSongs

UNLOCKMAN:SongIsLocked(song) will return an int between 0 to 15 inclusive, the
result of bitwise ORing powers-of-2 representing reasons a song could be locked,
defined in UnlockManager.h.

0 means the song is not locked for any reason, so let's ensure
UNLOCKMAN:SongIsLocked(song) returns 0 in ScreenSelectMusicCasual.  If not,
prune the song.