Fix PaneDisplay when Groovestats is enabled
The issue https://github.com/CrashCringle12/Simply-Love-SM5/issues/14 noted ScreenSelectMusic was broken in 4:3. The PR https://github.com/CrashCringle12/Simply-Love-SM5/pull/15 fixed the issue; however, it broke PaneDisplay when Groovestats is enabled (It was already broken most likely in 4:3, but it was now broken across board). This should fix the display across the board.

Next commit might provide some parity with other aspect ratios if I get to it as I'm notcing some things are a bit off.