Give players a way to see if a song is in their favorites in the  MusicWheel.
You can see if a song is in your favorites by the color of the Cursor.

Eventually I'd like to find a way so they can see which songs are in their favorites just by looking at the songs currently listed in the wheel. Presently you have to hover over the song to discover. It's a very small QOL thing though so pretty low priority.