Fix random bugs from merge/upstream, Adjust what is displayed in PaneDisplay
The current display is not favorable for students at PSU,
as they are much more interested in local leaderboard as global leaderboard is more distant/irrelevant to them.
They seem much more interested in seeing the club's history through leaderboards and feeling accomplished
reaching levels previous club members have. The global one I think feels far away and less of
a connection to it.
"Why does it have to keep showing me these 100%s. I get it I suck and they're good."
Kind of a unique thing for PSU maybe? Not sure, but I definitely want to keep always
showing Machine High Score. I'm currently experimenting with fading in and out between scores
So in this commit we're fading in out and between the EX and ITG variants of each score if applicable.

I think I will add a padcode to swap between which scores are displayed.